+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/to_struct.rs
@@ -1,10 +1,7 @@
-use crate::conversions::v3_0::to_rust_type::{
-    object_to_field_shapes, shape_schema_object_type, DefinitionShape, FieldShape, TypeShape,
-};
+use crate::conversions::v3_0::to_rust_type::{object_to_field_shapes, DefinitionShape, FieldShape};
 use crate::conversions::Result;
-use crate::targets::rust_type::{DataType, StructDef};
-use openapi_types::v3_0::{ReferenceObject, SchemaCase, SchemaFieldName, SchemaObject};
-use SchemaCase::{Reference, Schema};
+use crate::targets::rust_type::StructDef;
+use openapi_types::v3_0::{SchemaFieldName, SchemaObject};
 
 pub(super) fn to_struct(name: SchemaFieldName, object: SchemaObject) -> Result<DefinitionShape> {
     let field_shapes = object_to_field_shapes(object)?;
@@ -30,24 +27,4 @@ pub(super) fn to_struct(name: SchemaFieldName, object: SchemaObject) -> Result<D
         DefinitionShape::Fixed(def.into())
     };
     Ok(shape)
-}
-
-pub(super) fn shape_type(schema_case: SchemaCase) -> Result<TypeShape> {
-    match schema_case {
-        Schema(object) => shape_schema_object_type(*object),
-        Reference(object) => shape_schema_reference_type(object),
-    }
-}
-
-fn shape_schema_reference_type(object: ReferenceObject) -> Result<TypeShape> {
-    let type_name = match String::from(object) {
-        x if x.starts_with("#/components/schemas/") => {
-            // TODO: change location to relative paths by TypeShape::Ref
-            // if using "#/components/responses/" etc
-            x.replace("#/components/schemas/", "")
-        }
-        x => unimplemented!("not implemented: {x}"),
-    };
-
-    Ok(TypeShape::Fixed(DataType::Custom(type_name)))
 }
